+++ noodles-util/src/alignment.rs
@@ -1,4 +1,7 @@
 //! Alignment format utilities.
 
+#[cfg(feature = "async")]
+pub mod r#async;
+
 pub mod io;
 pub mod iter;