+++ noodles-util/src/alignment/async.rs
@@ -1,0 +1,3 @@
+//! Async alignment format utilities.
+
+pub mod io;